requirements.txt
Pillow==12.0.0
schedule==1.2.2
